コミット

+++ src/main/java/com/example/macaron/Sale.java
@@ -6,14 +6,14 @@ public class Sale {
     private final String name;
     private final String contents;
 
-    public Sale(int id, int store_id, String name, String contents) {
+    public Sale(Long id, int store_id, String name, String contents) {
         this.id = id;
         this.store_id = store_id;
         this.name = name;
         this.contents = contents;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
     public int getStore_id() {
