+++ src/main/java/com/example/macaron/StoreReview.java
@@ -1,14 +1,13 @@
 package com.example.macaron;
 
-
 public class StoreReview {
-    private final int id;
+    private final Long id;
     private final int store_id;
     private final String comments;
     private final boolean evaluation;
     private final String user_id;
 
-    public StoreReview(int id, int store_id, String comments, boolean evaluation, String user_id) {
+    public StoreReview(Long id, int store_id, String comments, boolean evaluation, String user_id) {
         this.id = id;
         this.store_id = store_id;
         this.comments = comments;
@@ -16,7 +15,7 @@ public class StoreReview {
         this.user_id = user_id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
     public int getStore_id() {
